server/requirements.txt
# --- Web & Framework ---
fastapi==0.115.0
uvicorn==0.30.6
starlette==0.38.6

# --- DB ---
SQLAlchemy==2.0.36

# --- Utils ---
python-dotenv==1.0.1
joblib==1.4.2

# --- Data stack ---
numpy==1.26.4
pandas==2.2.2

# --- Vector search ---
faiss-cpu==1.8.0.post1   # manylinux 휠 존재(Py3.11에서 안정)

# --- OpenAI ---
openai>=1.52,<2

# --- (옵션) 필요할 때만 ---
scikit-learn==1.5.1
httpx>=0.25,<0.28
psycopg2-binary>=2.9
